third_party/WebKit/LayoutTests/fast/css/implicit-attach-marking.html

<!DOCTYPE html>

<script>jsTestIsAsync = true;</script>
<script src="../../resources/js-test.js"></script>

<div id="root"></div>

<script>
onload = function() {
    requestAnimationFrame(function() {
        div = root.appendChild(document.createElement("div"));
        div.textContent = "Test that implicit attach marks all ancestors on insert.";
        shouldBeTrue("div.offsetTop > 0");
        finishJSTest();
    });
};
</script>
