third_party/WebKit/LayoutTests/fast/css/implicit-attach-marking-expected.txt

PASS div.offsetTop > 0 is true
PASS successfullyParsed is true

TEST COMPLETE

Test that implicit attach marks all ancestors on insert.
